requirements.txt
torch
torchvision
Pillow
matplotlib
wandb
jupyter